mysite/requirements.txt
Django==2.1.7
django-haystack==2.8.1
django-redis==4.10.0
jieba==0.39
Pillow==5.4.1
py3Fdfs==2.2.0
PyMySQL==0.9.3
pytz==2018.9
redis==3.2.1
uWSGI==2.0.18
Whoosh==2.7.4
